backend/requirements.txt
fastapi
uvicorn
sqlalchemy
geoalchemy2
alembic
psycopg2-binary
google-auth>=2.0.0
google-auth-httplib2
slowapi

fastapi-cache2>=0.2.0
asyncpg
pydantic
